refactor(theme): migrate and refactor type `TokenWithCommonCls` by `@ant-design/cssinjs-utils`
