+++ components/theme/internal.ts
@@ -29,7 +29,7 @@ import genComponentStyleHook, {
 import genPresetColor from './util/genPresetColor';
 import useResetIconStyle from './util/useResetIconStyle';
 
-export type { CSSUtil } from '@ant-design/cssinjs-utils';
+export type { CSSUtil, TokenWithCommonCls } from '@ant-design/cssinjs-utils';
 
 export { DesignTokenContext, defaultConfig } from './context';
 export {
